【版本】2.1.15【介绍】优化了ORM的selectOne的查询效率【功能】1、自动完成Activity、Fragment、Dialog和View中View的绑定 2、MVP的支持 3、企业级日志系统 4、TextUtils、NumberUtils、IoUtils、ImageUtils工具类等 5、ORM数据库关系映射框架 6、多种颜色值代码 7、万能AdapterView适配器 8、ShadeView滑动颜色渐变控件 9、水平Tab滚动条 10、多行单选按钮组 11、字母导航控件 12、可嵌套在ScrollView内部的ListView 13、雷达扫描控件

+++ jackknife/src/main/java/com/lwh/jackknife/orm/dao/Dao.java
@@ -44,7 +44,7 @@ public interface Dao<T extends OrmTable> {
 
     T selectOne(QueryBuilder builder);
 
-    int selectCount();
+    long selectCount();
 
-    int selectCount(QueryBuilder builder);
+    long selectCount(QueryBuilder builder);
 }
